Add JNI library for T265 odometry
This is intended to be used as a submodule for the main FTC repository
if you wish to compile librealsense from source. The CMake script will
output librealsense2.so and libt265.so to libs/armeabi-v7a in the main
repository, which will automatically be bundled by Gradle.

Defining an `externalNativeBuild` in the main repository is avoided
since everyone will need the header files to compile the JNI program.