Support getting the pose through JNI
Once the pipeline is started, it will push the frames to an internal
queue, which will drop them if we don't dequeue them in time. This way,
the queue always contains the newest frame for us to retrieve when we
call `getPoseMeters()` from Java.

The pose is returned as an instance of `Pose2d` (from ftclib).